+++ client/components/Routers.js
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { StyleSheet, Text, View, Button } from 'react-native';
 import Home from './Home';
 import PhotoScreen from './PhotoScreen'
-import Test from "./Test";
+import UploadPhoto from "./UploadPhoto";
 
 
 const Routes = () => {
@@ -18,12 +18,11 @@ const Routes = () => {
     return (
         <Router>
             {isLoggedIn ? (
-                <Scene key='root'>
+                <Stack key='root'>
                     <Scene key='Home' component={Home} initial={true} />
                     <Scene key='PhotoScreen' component={PhotoScreen} title={PhotoScreen}/>
-                    <Scene key='Test' component={Test} title={Test}/>
-
-                </Scene>
+                    <Scene key='UploadPhoto' component={UploadPhoto} title={UploadPhoto}/>
+                </Stack>
             ) : (
                 <Scene key='root'>
                     <Scene key='Home' component={Home} initial={true} />
