+++ package.json
@@ -29,7 +29,7 @@
     "pg": "^8.7.1",
     "react": "17.0.1",
     "react-dom": "17.0.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-44.0.0.tar.gz",
+    "react-native": "0.64.3",
     "react-native-dotenv": "^3.3.1",
     "react-native-router-flux": "^4.3.1",
     "react-native-web": "0.17.1",
